requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
argcomplete==3.5.1
attrs==24.2.0
bump2version==1.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
commitizen==3.31.0
decli==0.6.2
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
prompt-toolkit==3.0.36
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
questionary==2.0.1
termcolor==2.5.0
tomlkit==0.13.2
typing_extensions==4.12.2
wcwidth==0.2.13
yarl==1.17.2
